refactor in preparation for adding the SparkActionRun class
