+++ tron/core/jobrun.py
@@ -9,8 +9,8 @@ from collections import deque
 import tron.metrics as metrics
 from tron import command_context
 from tron import node
-from tron.core.actionrun import ActionRun
-from tron.core.actionrun import ActionRunFactory
+from tron.core.actionrun.base import ActionRun
+from tron.core.actionrun.base import ActionRunFactory
 from tron.serialize import filehandler
 from tron.utils import maybe_decode
 from tron.utils import next_or_none
